requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Memory & Embeddings
sentence-transformers==2.2.2
torch==2.5.1+cpu
faiss-cpu==1.7.4
numpy==1.24.3

# Optional: for better performance
python-multipart==0.0.6
